requirements.txt
ddddocr
Flask
flask-cors
requests
numpy
opencv-python-headless
Pillow
